Definition of the PhysicalSet object and the default metaprogramming initialization, model add and deletion.
The idea is to have the proper mixture form for the model, either
one model, a vector of model, or in the case of binary diffusion, a
matrix of models. This is taken care of by the structure SetOrEquation<Model>,
which will produce one pointer, a vector or a matrix of pointers.

Then all the init, add_model and destruction methods are general methods,
specialized through tags.

The tags is a collection of tags, defined to take care of every possible
situation, defaulted and derived for every case to limit writing.