not_implemented better taken care of:
In a very general way, only the change_file(const std::string & filename) and the
bool initialize() are critical.

The capacity to change file cannot be
specific to this or that parser type, because it is not parsing-related.
There is not format issue linked to it, we cannot afford to have a
parser that could be called to change a file and just throw an error.
The program would not notice because the old file is valid and
would be used.

The initialize for the kinetics is more of an historical legacy.  It is
used to avoid any useless operation if no kinetics information is
present in the file (l. 218 of read_reaction_set_data.h).  It is
easier for error-finding purpose to do it separately instead of
putting it in the bool reaction() method (l. 288), because this
will just consider that the reactions are read, and not that there is
no kinetics information in the file. Who knows what may come?  An easy
way out of it is to test the number of reactions before exiting,
maybe go this way?

Now every parser has different capabilities, and this is much easier
to consider that any call to any non-existing parsing capability is
equivalent to any other one, thus they all should be treated in the
same way, thus in the base object.

As of now, it's a warning, make it an error?