Update of all the kinetics model tests.
Now all the tests are standardized, testing both KineticsConditions and
StateType for the temperature. All the vectorizations are also tested.

PB: This sat around for awhile (sorry @SylvainPlessis) and I cherry
picked this. I manually updated the new CppUnit Arrhenius test to use
the KineticConditions in addition to the temperature.