MixtureConductivity version on build_eucken
Note that we're not trying to be clever and use traits/policy
to figure out that Eucken doesn't need to store n_species worth
of objects. But, because Eucken doesn't store any (very little now,
will be deprecated) state, the compiler ought to do the right thing
and not destroy our cache.