Remove ThermoEvaluator from TransportMixture
This is a major diff, but it was unavoidable as many objects/funcs
were templated on ThermoEvaluator because of TransportMixture.

Previously, TransportMixture was templated on a ThermoEvaluator type
and a const& was cached in the TransportMixture. At the time, this
was threadsafe because the type being used for ThermoEvaluator didn't
store any state that would've changed at thread time. However, this
violated our Mixture/Evaluator paradigm and could've led to unexpected
breakage if someone altered the ThermoEvaluator to cache thread-changing
state and didn't pay attention. Further, we really didn't need the
ThermoEvaluator to be cached in the TransportMixture anyway - it
was only there because of trying to construct conductvity objects
through the PhysicalSet infrastructure.