Add MixtureDiffusionBase
We're going to use this as a base class to define the interface
for the Mixture*Diffusion models. This way, we enforce the interface
and the user supply "empty error functions" for methods that are
inappropriately called when static_assert is not available (which
should be guarding upstream in WilkeTransportEvaulator).