Actually try calling the species method.
The hope was to have a static_assert in here, but it seems the compiler
is going into this function when I didn't think it was supposed to.
Help me @roystgnr!

Turns out we can't use static_assert that way.

Even though the function call will be optimized out in the
WilkeTransportEvaluator, the function still gets instantiated
first and the instantiation will fail.