Antioch no longer header-only
Have moved the implementation of get_antioch_version() to
utilities/src/antioch_version.C. The plan is to move all things that
can be templated on a "scalar" type to a .C file, but things that really
need a "CoeffType", i.e. things that could be expressions, we leave in
the header.

Need to link tests against libantioch.la