Logic fix in static_assert in WilkeTransportEvaluator
WilkeTransportEvaluator::mu_and_k_and_D can also handle the case of
thermal conductivity models that don't require diffusion together
with species diffusion models. The static_assert did not reflect
this.

We didn't hit this because of inadequate coverage in the
wilke_transport_unit test. This will be fixed subsequently.