Add vector versions of GSLSpliner tests
Within the CppUnit framework. Instead of splitting up the tests further
I just put all the different vector types in there knowing that
we'll see if Eigen, VexCL, etc. are skipped or not from other tests
so we'll know implicitly whehter they're being skipped in the GSL tests
too or not.