Two expected fail tests are added for XMLParser
  - one to test the name="k0" multiple attribute (fail_parsing_xml_1)
  - another to test the value other than "k0" restriction (fail_parsing_xml_2)

We use the same program than for the regular xml test. The xml files
only inclue the reaction to trigger the failure.
Some update of the test had to be made: it requires three input files,
originally, only the path to the folder was given, the names of the
files were hard coded in the code.  Now all three files are given
explicitely in the sh file.