Adding NASACurveFitBase
Collecting common code and data structures. We're going to be
refactoring NASA9 vs. CEA in the constructor. Further, I want
to remove the assumption on fixed intervals for NASA9. This is
going to facilitate that. Note there was an error in the code
for interval() copied from NASA7 that was fixed here (only showed
up in the vec tests).