Always initialize() XMLParser at construction
All the initialize() function is doing is setting up the pointers
to the different XMLNodes, so we should just do that instead of making
the user do it. Long term, we just need to remove the initialize() bit
as a public member, but that will be for a later day.