Fix XML thermo parsing for NASA9
This diff is ugly, but it's unavoidable. Previously, the sectioning was
incorrect. Furthermore, there was some duplicate code for effectively
doing the first iteration of the while loop over each of the intervals
for the coefficient data. This fixes all of those issues. However, it's
still hardcoded to support NASA9 only. Future commits will generalize to
parse NASA7 (keyword "NASA") as well as NASA9.