Add CppUnit test for NASA9 XML parsing
This doesn't yet test parsing the species-we-want-to-use from the XML
file, but it does test fairly completely the parsed coefficients from
the species present in the chemical mixture. We use the shared
nasa9_thermo.xml as the test file so we're actually testing (for N2 and
NO2 at least) the file that we're shipping.