Don't distribute license tools
It was annoying that the license tool got run in a distributed
tar ball and would probably confuse users. This removes it from
the distributed tarball. But we have to then check if it's there
if we're going to run it so that we don't get an error in the
tar ball (or different confusing messages).