The logic between Kooij / Arrhenius badly represented was faulty:
The problem lied in the difference of nature of the test
between the power parameter and the associated Tref parameter.
The power parameter relied on the parser finding one and whether the
chemical process was a falloff or not. On the other hand, Tref relied
on the type of kinetics model. So an Arrhenius coded as a Kooij equivalent
to an Arrhenius (power parameter == 0) would be recognized as an Arrhenius,
but, within a falloff, would accept the zero power parameter.

It leads to a data vector that contains 4 parameters instead of 3 for Arrhenius
or 5 for Kooij.