MicroThermoBase is now a CRTP base class
cv_vib() and cv_vib_over_R APIs have been moved to MicroThermoBase
and there are now cv_vib_impl and cv_vib_over_R_impl in each of
StatMechThermodynamics and IdealGasMicroThermo. I've made the *impl
methods private so we need to friend the base class in that case.

This gets us to the point where we have a uniform interface for
everything that was in IdealGasMicroThermo. We now need to extend
IdealGasMicroThermo (and MicroThermoBase) to get a uniform API
for all macro thermo and micro thermo the user might want to use
between the two MicroThermoBase subclasses.