Rename MicroThermoBase to MacroMicroThermoBase
Since we're defining an API for both macro and micro thermo to be
used by the user, this name makes more sense for the class, I think.
This still may not be the right final design (e.g. we want want to
separate macro and mirco thermo), but we'll go with this for now.