Fix remove_newline_from_strings for multiple "\n"
If the vector of strings contains newline characters, the algorithm that
was there was broken because it started referencing a null iterator. I
caught this trying to parse GRI3 XML thermo coeffs, so unit test
following this will mimic that trigger.

This was actually an epic fail on my part because the fix was doing
exactly the same thing 5 lines below for cleaning out each element.
*facepalm*