Add test_cv_vib_over_R for StatMechThermo
Note that this isn't standing up to testing in long double precision.
My guess is the exp implementation, but that's just a guess. And seeing
that we'll probably hit the same with cv_el, I just created a second
block of testing classes that won't get hit with long double.