Add logic in XMLParser for specific thermo/reaction data
If there is a phase block, we check the speciesArray/datasrc and
reactionArray/datasrc attributes (error out if they're not there)
and then set the apporiate XML pointers to the right places in the
file. Note we adhere to what I believe is Cantera's convention of
putting a '#' at the beginning of the name to refer to internal
blocks (as opposed to external filenames, like elements.xml).