Fix grabbing wrong mixture
This was sneaky and I only noticed where there was an erroant
message of opening the default transport file. We wanted a
TransportMixture, but we grabbed a ChemicalMixture (probably a
copy/paste bug). But TransportMixture has a bunch of default
arguments so a ChemicalMixture was enough to trigger construction
of a TransportMixture with default things. This would've been
frustrating once we tried to do this with species that weren't
in the default transport data file.